+++ src/Trade360SDK.Common.Entities/Entities/OutrightLeague/OutrightLeagueFixture.cs
@@ -24,7 +24,7 @@ namespace Trade360SDK.Common.Entities.OutrightLeague
 
         public IEnumerable<NameValuePair>? ExtraData { get; set; }
         
-        public DateTime EndDate { get; set; }
+        public DateTime? EndDate { get; set; }
         
         public IdNamePair? Season { get; set; }
     }